% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/weighting.R
\name{AdjustDataToReflectWeights}
\alias{AdjustDataToReflectWeights}
\title{Adjust data to reflect weights}
\usage{
AdjustDataToReflectWeights(data, weights, seed = 123)
}
\arguments{
\item{data}{A \code{\link{data.frame}}.}

\item{weights}{The sampling or replication weights.}

\item{seed}{The seed used in random number generation.}
}
\description{
Creates a new \code{\link{data.frame}} to reflect weights.
}
\details{
In situations where an algorithm does not accomodate weights, this
function modifies the \code{\link{data.frame}} by either:
(A) stretching it out, where the the weights are integers, or
(B) resampling to create a new bootstrapped \code{\link{data.frame}},
where the \code{weights} are proportional to the probability of
selection.
When creating the bootstrap sample, the sample size is whichever is greatest
of the rounded sum and 1.
}

