% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/weighting.R
\name{AdustDataToReflectWeights}
\alias{AdustDataToReflectWeights}
\title{\code{AdustDataToReflectWeights} Adjust data to reflect weights.
Creates a new \code{\link{data.frame}} to reflect weights.}
\usage{
AdustDataToReflectWeights(data, weights, seed = 123)
}
\arguments{
\item{data}{A \code{\link{data.frame}}.}

\item{weights}{The sampling or replication weights.}

\item{see}{The seed used in random number generation.}
}
\description{
\code{AdustDataToReflectWeights} Adjust data to reflect weights.
Creates a new \code{\link{data.frame}} to reflect weights.
}
\details{
In situations where an algorithm does not accomodate weights, this
function modifies the \code{\link{data.frame}} by either:
(A) stretching it out, where the the weights are integers, or
(B) resampling to create a new bootstrapped  \code{\link{data.frame}},
where the  \code{weights} are proportional to the probability of
selection.
When creating the bootstrap sample, the sample size is whichever is greatest
of the  rounded sum and 1.
}

