% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingdata.R
\name{AnyNA}
\alias{AnyNA}
\title{\code{AnyNA}}
\usage{
AnyNA(data, formula = NULL)
}
\arguments{
\item{data}{A \code{\link{data.frame}}.}

\item{formula}{A no optional \code{\link{formula}}. Variables not listed in a formula are excluded from the evaluation.}
}
\description{
Checks to see if the are any missing valus
}

