% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{BinaryLogit}
\alias{BinaryLogit}
\title{Binary Logistic Regression}
\usage{
BinaryLogit(formula, data, weights = NULL, subset = NULL, ...)
}
\arguments{
\item{...}{Additional arguments affecting the goodness-of-fit displayed.}
}
\description{
Binary Logistic Regression
}

