% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressiondiagnostics.R
\name{BreuschPagan}
\alias{BreuschPagan}
\title{\code{BreuschPagan} Breusch-Pagan test for non-constant variance.}
\usage{
BreuschPagan(Regression.object, show.warnings = TRUE)
}
\arguments{
\item{show.warnings}{Returns a warning if the sample size is less than 4.}

\item{mod}{An object of class \code{\link{Regression}}.}
}
\description{
\code{BreuschPagan} Breusch-Pagan test for non-constant variance.
}
\details{
Weights are ignored when conducting the test.
}

