% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/missingdata.R
\name{CleanSubset}
\alias{CleanSubset}
\title{\code{CleanSubset}
Takes a QSubset variable and turns it into a logical vector with no missing values}
\usage{
CleanSubset(subset, n.total)
}
\arguments{
\item{n.total}{The total numbe of observations.}

\item{q.subset}{A QSubset variable from Displayr or Q.}
}
\description{
\code{CleanSubset}
Takes a QSubset variable and turns it into a logical vector with no missing values
}

