% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingdata.R
\name{CleanWeights}
\alias{CleanWeights}
\title{\code{CleanWeights}
Takes a QSubset variable and turns it into a logical vector with no missing values}
\usage{
CleanWeights(weights)
}
\arguments{
\item{weights}{A weights variable from Displayr or Q (i.e., \code{QcalibratedWeight}
or \code{QPopulationWeight}).}
}
\description{
\code{CleanWeights}
Takes a QSubset variable and turns it into a logical vector with no missing values
}

