% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/factoranalysis.R
\name{ComponentPlot}
\alias{ComponentPlot}
\title{\code{ComponentPlot}}
\usage{
ComponentPlot(x)
}
\arguments{
\item{x}{An object of class \code{flipFactorAnalysis}.}
}
\description{
Create a scatter plot showing the loadings of each variable on the first two principal components.
}

