% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressiondiagnostics.R
\name{ConfusionMatrixFromVariables}
\alias{ConfusionMatrixFromVariables}
\title{\code{ConfusionMatrixFromVariables}}
\usage{
ConfusionMatrixFromVariables(observed, predicted, subset = NULL,
  weights = NULL)
}
\arguments{
\item{observed}{A \code{factor}.}

\item{predicted}{A \code{factor}.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process, or, the name of a variable in \code{data}. It
may not be an expression. \code{subset} may not}

\item{weights}{An optional vector of sampling weights, or, the name or, the
name of a variable in \code{data}. It may not be an expression.}
}
\description{
\code{ConfusionMatrixFromVariables}
}
\details{
A contingency table showing the observed versus predicted values from a model.
}

