% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressiondiagnostics.R
\name{ConfusionMatrixFromVariablesLinear}
\alias{ConfusionMatrixFromVariablesLinear}
\title{\code{ConfusionMatrixFromVariablesLinear}}
\usage{
ConfusionMatrixFromVariablesLinear(observed, predicted, subset = NULL,
  weights = NULL)
}
\arguments{
\item{observed}{A \code{factor}.}

\item{predicted}{A \code{factor}.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process, or, the name of a variable in \code{data}. It
may not be an expression. \code{subset} may not}

\item{weights}{An optional vector of sampling weights, or, the name or, the
name of a variable in \code{data}. It may not be an expression.}
}
\description{
\code{ConfusionMatrixFromVariablesLinear}
}
\details{
A contingency table showing the observed versus predicted values from a model, for linear models.
Prediced values are assigned the value of the closest observed value.
}

