% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressiondiagnostics.R
\name{CooksDistance}
\alias{CooksDistance}
\title{\code{CooksDistance}}
\usage{
CooksDistance(model)
}
\arguments{
\item{model}{A 'Regression'  model.}
}
\description{
\code{CooksDistance}
}
\details{
Computes Cook's distance and a threshold value.
}

