% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeplearning.R
\name{dataToNumeric}
\alias{dataToNumeric}
\title{Takes a data frame as input as converts output to a numeric matrix
Unlike data.matrix, factors are converted to an indicator matrix
We don't want to remove missing levels from prediction data}
\usage{
dataToNumeric(df)
}
\description{
Takes a data frame as input as converts output to a numeric matrix
Unlike data.matrix, factors are converted to an indicator matrix
We don't want to remove missing levels from prediction data
}

