% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressiondiagnostics.R
\name{diagnosticTestFromCar}
\alias{diagnosticTestFromCar}
\title{#' @export
marginalModelPlots.Regression <- function(model, ...)
{
    checkAcceptableModel(model, c("glm","lm"), "residualPlots")
    diagnosticTestFromCar(model, "marginalModelPlots", ...)
}}
\usage{
diagnosticTestFromCar(x, diagnostic, ...)
}
\description{
#' @export
marginalModelPlots.Regression <- function(model, ...)
{
    checkAcceptableModel(model, c("glm","lm"), "residualPlots")
    diagnosticTestFromCar(model, "marginalModelPlots", ...)
}
}

