% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{DichotomizeFactor}
\alias{DichotomizeFactor}
\title{\code{DichotomizeFactor} Converts a list of variable or data frames into a
data.frame.}
\usage{
DichotomizeFactor(variable, cutoff = 0.5, warning = FALSE,
  variable.name = deparse(substitute(variable)))
}
\arguments{
\item{variable}{A variable in a DataSet or data.frame.}

\item{cutoff}{The cutoff point to split the variable into.}

\item{warning}{If TRUE, raise a warning showing the new levels.}

\item{variable.name}{An alternate name to show instead of the deparsed
variable name.}
}
\description{
\code{DichotomizeFactor} Converts a list of variable or data frames into a
data.frame.
}

