% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{DiscriminantVariables}
\alias{DiscriminantVariables}
\title{\code{DiscriminantVariables}}
\usage{
DiscriminantVariables(x)
}
\arguments{
\item{x}{A \code{LDA} object.}
}
\description{
Computes discriminant variables for the entire sample passed into the original analysis (including missing and filtered values).
}
