% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressiondiagnostics.R
\name{DurbinWatson}
\alias{DurbinWatson}
\title{\code{DurbinWatson}}
\usage{
DurbinWatson(model, n.permutations = 1000)
}
\arguments{
\item{n.permutations}{Number of permutations used in computing the p-value.}

\item{residuals}{Residuals from a model.}
}
\description{
\code{DurbinWatson}
}
\details{
Computes the Durbin-Watson statistic. A permutation test is used for
computing the p-value. Tests to a lag of 1. Two-sided.
Durbin, J.; Watson, G. S. (1950). "Testing for Serial Correlation in Least Squares Regression, I". Biometrika 37 (3–4): 409–428.
}

