% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatting.R
\name{Equation}
\alias{Equation}
\title{\code{Equation} Formats a model as an equation (string), for
description purposes.}
\usage{
Equation(object)
}
\arguments{
\item{object}{The model.}
}
\description{
\code{Equation} Formats a model as an equation (string), for
description purposes.
}

