% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingdata.R
\name{ErrorIfMissingDataFound}
\alias{ErrorIfMissingDataFound}
\title{\code{ErrorIfMissingDataFound}}
\usage{
ErrorIfMissingDataFound(data)
}
\arguments{
\item{data}{A \code{\link{data.frame}}.
# @examples
# df = data.frame(x = c(NA, 1), y = 1:2)
# ErrorIfMissingDataFound(df)}
}
\description{
This is a wrapper for \code{\link{na.fail}}
}

