% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/missingdata.R
\name{ExcludeCasesWithAnyMissingData}
\alias{ExcludeCasesWithAnyMissingData}
\title{\code{ExcludeCasesWithAnyMissingData}}
\usage{
ExcludeCasesWithAnyMissingData(data)
}
\arguments{
\item{data}{A \code{\link{data.frame}}.}
}
\description{
This is a wrapper for \code{\link{na.omit}}
}
\examples{
df <- data.frame(x = c(NA, 1), y = 1:2)
ExcludeCasesWithAnyMissingData(df)
}

