% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/missingdata.R
\name{ExcludeCasesWithCompletelyMissingData}
\alias{ExcludeCasesWithCompletelyMissingData}
\title{\code{ExcludeCasesWithCompletelyMissingData}}
\usage{
ExcludeCasesWithCompletelyMissingData(data)
}
\arguments{
\item{data}{A \code{\link{data.frame}}.}
}
\description{
Create a copy of a data frame where any cases which have only missing values
are removed.
}
\examples{
my.df <- data.frame("A" = c(1, 2, 3, 4, NA), "B" = c(NA, 1, 2, 3, NA), "C" = c(NA, NA, 1, 2, NA))
ExcludeCasesWithCompletelyMissingData(my.df)
}

