% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/factoranalysis.R
\name{FactorAnalysis}
\alias{FactorAnalysis}
\title{\code{FactorAnalysis}}
\usage{
FactorAnalysis(data, weights = NULL, subset = NULL,
  missing = "Exclude cases with missing data", use.correlation = TRUE,
  type = "PCA", rotation = "none", n.factors = 1,
  sort.coefficients.by.size = FALSE, suppress.small.coefficients = FALSE,
  min.display.loading.value = 0.1)
}
\arguments{
\item{data}{A data framer with numeric columns which contains the data to be analyzed.}

\item{weights}{A numeric vector containing the weight for each case in data.}

\item{subset}{A logical vector which describes the subset of \code{data} to be analyzed.}

\item{missing}{A string specifiying what to do when the \code{data} contains missing values.
The valid options are \code{"Error if missing data found"}, \code{"Exclude cases
with missing data"}, \code{"Use partial data"}, and \code{"Imputation"}.}

\item{use.correlation}{A logical value specifying whether to use the correlation matrix (\code{TRUE}),
or the covariance matrix (\code{FALSE}).}

\item{type}{A string specifying the type of analysis to do. Valid options are \code{"PCA"},
\code{"Unweighted least squares"}, \code{"Generalized least squares"}, and
\code{"Maximum likelihood"}.}

\item{rotation}{A string specifying the type of rotation to be used. Valid options are}
}
\description{
Calculate a Factor Analysis or Principal Component Analysis
}

