% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{FactorToIndicators}
\alias{FactorToIndicators}
\title{Converts a factor into an indicator matrix.}
\usage{
FactorToIndicators(variable, variable.name = deparse(substitute(variable)))
}
\arguments{
\item{variable}{A variable in a \code{DataSet} or \code{data.frame}.}

\item{variable.name}{The name of the variable.}
}
\description{
Converts a factor into an indicator matrix.
}

