% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transformations.R
\name{FactorToNumeric}
\alias{FactorToNumeric}
\title{Converts a an ordered factor into a numeric variable}
\usage{
FactorToNumeric(variable, variable.name = deparse(substitute(variable)))
}
\arguments{
\item{variable}{A variable in a \code{DataSet} or \code{data.frame}.}

\item{variable.name}{The name of the variable.}
}
\description{
Converts a an ordered factor into a numeric variable
}

