% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodnessoffit.R
\name{FittedAndObserved}
\alias{FittedAndObserved}
\title{Extracts fitted and observed.
\code{FittedAndObserved} extracts vectors of fitted (or predicted) and observed values from an object for use in computing
\code{GoodnessOfFit}.}
\usage{
FittedAndObserved(object)
}
\arguments{
\item{object}{An object for which a summary is desired..}
}
\value{
fitted Fitted values from the object.

observed Observed values from the object.
}
\description{
Extracts fitted and observed.
\code{FittedAndObserved} extracts vectors of fitted (or predicted) and observed values from an object for use in computing
\code{GoodnessOfFit}.
}

