% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{FormatAsPValue}
\alias{FormatAsPValue}
\title{\code{FormatAsPValue}
Formats p relatively nicely, ensuring that values greater than 0.05
are not rounded to 0.05, and numbers greater than 0 are not rounded to 0,
unless is really close to 0.}
\usage{
FormatAsPValue(p, p.cutoff = 0.05, ...)
}
\arguments{
\item{p}{The number(s)}

\item{p.cutoff}{TODO}

\item{...}{Additional arguments that are passed to \code{\link{formatC}}.}
}
\description{
\code{FormatAsPValue}
Formats p relatively nicely, ensuring that values greater than 0.05
are not rounded to 0.05, and numbers greater than 0 are not rounded to 0,
unless is really close to 0.
}

