% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatting.R
\name{FormatAsReal}
\alias{FormatAsReal}
\title{\code{FormatAsReal} Formats real numbers nicely, as if they are to presented as a an average.}
\usage{
FormatAsReal(x, digits = 2, format = "fg", ...)
}
\arguments{
\item{x}{The number(s)}

\item{digits}{Number of significant digits}

\item{format}{See \code{\link{formatC}}.}

\item{...}{Additional arguments that are passed to  \code{\link{formatC}}.}
}
\description{
\code{FormatAsReal} Formats real numbers nicely, as if they are to presented as a an average.
}

