% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/factoranalysis.R
\name{GenerateScoresFromFactorAnalysis}
\alias{GenerateScoresFromFactorAnalysis}
\title{\code{GenerateScoresFromFactorAnalysis}}
\usage{
GenerateScoresFromFactorAnalysis(factor.analysis.object,
  method = "Regression")
}
\arguments{
\item{factor.analysis.object}{The results of a factor analysis or principal component analysis
that was done using \link{\code{FactorAnlysis}}. This provides the
loadings, the raw data to transform, as well as wieghts and subset
information.}

\item{method}{A string describing the method to use to obtain the scores. The allowed options are
              \code{Regression}, \code{Bartlett}, and \code{Anderson-Rubin}, and these map to the
              options used by package psych \code{Thurstone}, \code{Bartlett}, and \code{Anderson}
              respectively.}
}
\value{
A data frame with the same dimensions as the data which was originally supplied to
         \link{\code{FactorAnalysis}}. Any
}
\description{
Generates a set of scores from a factor analysis or principal component analysis,
              which is a new set of data corresponding to the principal components or factors.
}
\details{
This function provides a wrapper for the function \code{factor.scores} from package psych.
         It adds functionality to handle the presence of weights. Where weights are specified
         the input data is standardized using the weighted standard deviation and mean. The data that
         is used to generate the scores is the same as that used to generate the factor analysis
         or PCA. That is, the original subset is isued, if imputation was specified originally, then
         the imputed data is used, otherwise all cases which are not completely missing are used.
}

