% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodnessoffit.R
\name{GoodnessOfFitInternal}
\alias{GoodnessOfFitInternal}
\title{Packages up the object goodness of fit,
\code{GoodnessOfFitInternal} puts together the various bits of \code{GoodnessOfFit}.  These should usually not
be used directly unless by experienced users.}
\usage{
GoodnessOfFitInternal(value, description, call)
}
\arguments{
\item{value}{The computed goodness-of-fit.}

\item{description}{Text elements used to construct the print statment.}

\item{call}{The original call used to create the object for which goodness-of-fit is being computed.}

\item{...}{Additional arguments affecting the goodness-of-fit displayed.}
}
\value{
A \code{\link{list}} with components

value The computed goodness-of-fit.

description Text elements used to construct the print statment.

call The original call used to create the object for which goodness-of-fit is being computed.
}
\description{
Packages up the object goodness of fit,
\code{GoodnessOfFitInternal} puts together the various bits of \code{GoodnessOfFit}.  These should usually not
be used directly unless by experienced users.
}

