% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientboost.R
\name{GradientBoost}
\alias{GradientBoost}
\title{\code{GradientBoost}}
\usage{
GradientBoost(formula, data = NULL, subset = NULL, weights = NULL,
  output = "Accuracy", missing = "Exclude cases with missing data",
  booster = "gbtree", search.rounds = 0, seed = 12321,
  show.labels = FALSE)
}
\arguments{
\item{formula}{A formula of the form \code{groups ~ x1 + x2 + ...}
That is, the response is the grouping factor and the right hand side
specifies the (non-factor) discriminators, and any transformations, interactions,
or other non-additive operators will be ignored.}

\item{data}{A \code{\link{data.frame}} from which variables specified
in formula are preferentially to be taken.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process, or, the name of a variable in \code{data}. It
may not be an expression.}

\item{weights}{An optional vector of sampling weights, or the
name of a variable in \code{data}. It may not be an expression.}

\item{output}{One of \code{"Accuracy"}, \code{"Importance"},
\code{"Prediction-Accuracy Table"} or \code{"Detail"}.}

\item{missing}{How missing data is to be treated. Options:
\code{"Error if missing data"},
\code{"Exclude cases with missing data"}.}

\item{booster}{Whether the underlying model is a tree or linear regression. Options:
\code{"gbtree"},
\code{"gblinear"}.}

\item{search.rounds}{The number of sets of random paramaters to search in order to tune the model.
Zero implies using the default paramaters.}

\item{seed}{The random number seed used in imputation.}

\item{show.labels}{Shows the variable labels, as opposed to the labels, in the outputs, where a
variables label is an attribute (e.g., attr(foo, "label")).}
}
\description{
GradientBoost
}
