% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressiondiagnostics.R
\name{HatValues}
\alias{HatValues}
\title{\code{HatValues}}
\usage{
HatValues(model)
}
\arguments{
\item{model}{A 'Regression' model.}
}
\description{
\code{HatValues}
}
\details{
Computes hat values and a threshold value.
}

