% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineardiscriminantanalysis.R
\name{LDA.fit}
\alias{LDA.fit}
\title{\code{LDA.fit} Fits linear discriminant analysis models.}
\usage{
LDA.fit(x, grouping, prior = proportions, tol = 1e-04,
  method = c("moment", "mle"), weights = NULL, CV = FALSE, nu = 5,
  labels = NULL, functions.output = FALSE, ...)
}
\arguments{
\item{x}{A matrix or data frame of explanatory variables.}

\item{grouping}{A variable containg the group memberships (i.e., to be
predicted or explained by x).}

\item{prior}{The assumed probability of each value of y occurring in the
population.  By default this is set to "observed" and the value is computed
based on the observed data.  If set to "constant" the prior will be set to
be equal for each group (this is the default in SPSS).  Alternatively, a
vector of probabilities can be provided.}

\item{tol}{Tolerance to decide if a matrix is singular.}

\item{method}{The method used to estimate the variance; either \code{"moment"} for
the method of moments or \code{"mle"} for maximum likelihood estimaion.}

\item{weights}{An optional vector of sampling weights.}

\item{CV}{Not used.}

\item{nu}{the number of left singular vectors to be computed.}

\item{labels}{The labels of the predictor variables.}

\item{functions.output}{Logical; whether the discriminant functions are the
required output of \code{\link{LDA}}.}

\item{...}{Additional arguments.}
}
\description{
\code{LDA.fit} Fits linear discriminant analysis models.
}
\details{
This is a wrapper for MASS::lda and MASS::qda.
  #### Linear discriminant analysis #####
  \url{http://www.ats.ucla.edu/stat/spss/output/SPSS_discrim.htm}
}
