% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineardiscriminantanalysis.R
\name{LDA}
\alias{LDA}
\title{\code{LDA}}
\usage{
LDA(formula, data = NULL, subset = NULL, weights = NULL,
  prior = "Equal", missing = "Exclude cases with missing data",
  output = "Means", outcome.color = "#5B9BD5",
  predictors.color = "#ED7D31", variance = "moment", seed = 12321,
  statistical.assumptions, auxiliary.data = NULL, show.labels = FALSE, ...)
}
\arguments{
\item{formula}{A formula of the form \code{groups ~ x1 + x2 + ...}
That is, the response is the grouping factor and the right hand side
specifies the (non-factor) discriminators, and any transformations, interactions,
or other non-additive operators will be ignored.
transformations nor}

\item{data}{A \code{\link{data.frame}} from which variables specified
in formula are preferentially to be taken.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process, or, the name of a variable in \code{data}. It
may not be an expression. \code{subset} may not}

\item{weights}{An optional vector of sampling weights, or, the name or, the
name of a variable in \code{data}. It may not be an expression.}

\item{prior}{The assumed probability of each value of y occurring in the
population.  By default this is set to "observed" and the value is computed
based on the observed data.  If set to "equal" the prior will be set to
be equal for each group (this is the default in SPSS).  Alternatively, a
vector of probabilities can be provided.}

\item{missing}{How missing data is to be treated in the regression. Options:
\code{"Error if missing data"}
\code{"Exclude cases with missing data"}
\code{"Imputation (replace missing values with estimates)"}}

\item{output}{One of \code{"Means"}, \code{"Confusion matrix"}, or \code{"Detail"}.
\code{"Scatterplot"}, and \code{"Moonplot"}.}

\item{outcome.color}{Color used to display centroids in Scatterplot output.}

\item{predictors.color}{Color used to display variable correlations in Scatterplot output.}

\item{variance}{The method used to estimate the variance; either \code{"moment"} for
the method of moments or \code{"mle"} for maximum likelihood estimaion.}

\item{seed}{The random number seed used in imputation.}

\item{statistical.assumptions}{A Statistical Assumptions object.}

\item{auxiliary.data}{A \code{\link{data.frame}} containing additional variables
to be used in imputation (if required). While adding more variables will improve
the quality of the imputation, it will dramatically slow down the time to estimate.
Factors and Character variables with a large number of categories should not be included,
as they will both slow down the data and are unlikely to be useful}

\item{show.labels}{Shows the variable labels, as opposed to the labels, in the outputs, where a
variables label is an attribute (e.g., attr(foo, "label")).}

\item{...}{Additional argments to be past to  \code{LDA.formula}.}
}
\description{
Linear Discriminant Analysis
}
\details{
"Imputation (replace missing values with estimates): All selected
  outcome and predictor variables are included in the imputation, along with
  all \code{auxiliary.data}, excluding cases that are excluded via subset or
   have invalid weights, but including
  cases with missing values of the outcome variable.
  Then, cases with missing values in the outcome variable are excluded from
  the analysis (von Hippel 2007). See \code{\link[flipImputation]{Imputation}}.
}
\references{
von Hippel, Paul T. 2007. "Regression With Missing Y's: An
  Improved Strategy for Analyzing Multiply Imputed Data." Sociological
  Methodology 37:83-117. White, H. (1980), A heteroskedastic-consistent
  covariance matrix estimator and a direct test of heteroskedasticity.
  Econometrica, 48, 817-838. Long, J. S. and Ervin, L. H. (2000). Using
  heteroscedasticity consistent standard errors in the linear regression
  model. The American Statistician, 54(3): 217-224.
}

