% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lineardiscriminantanalysis.R
\name{LinearDiscriminantAnalysis}
\alias{LinearDiscriminantAnalysis}
\title{LinearDiscriminantAnalysis
\code{LinearDiscriminantAnalysis} Fits linear discriminant analysis models.}
\usage{
LinearDiscriminantAnalysis(x, grouping, prior = proportions, tol = 1e-04,
  method = c("moment", "mle"), weights = rep(1, n.cases), CV = FALSE,
  nu = 5, ...)
}
\arguments{
\item{x}{A matrix or data frame of explanatory variables.}

\item{prior}{The assumed probability of each value of y occurring in the population.  By default this is set to "observed" and
the value is computed based on the observed data.  If set to "constant" the prior will be set to be equal for each group (this is the
default in SPSS).  Alternatively, a vector of probabilities can be provided.}

\item{...}{Additional arguments.}

\item{y}{A variable containg the group memberships (i.e., to be predicted or explained by x).}
}
\description{
LinearDiscriminantAnalysis
\code{LinearDiscriminantAnalysis} Fits linear discriminant analysis models.
}
\details{
This is a wrapper for MASS::lda and MASS::qda.

#### Linear discriminant analysis #####
}

