% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regression.R
\name{LinearRegression}
\alias{LinearRegression}
\title{Object Goodness-of-Fit Plot
\code{Linear Regression} A generic function used to produce plots showing outliers in the data from
a model object.  The function invokes particular \code{\link{methods}}
which depend on the \code{\link{class}} of the first argument.}
\usage{
LinearRegression(formula, data, ...)
}
\arguments{
\item{...}{Additional arguments affecting the goodness-of-fit displayed.}

\item{object}{An object for which a summary is desired..}
}
\description{
reports the goodness-of-fit of an object.
}

