% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regression.R
\name{linearRegressionFromCorrelations}
\alias{linearRegressionFromCorrelations}
\title{#Quasi-Binomial Regression
#
#}
\usage{
linearRegressionFromCorrelations(formula, data, subset = NULL,
  weights = NULL, outcome.variable, outcome.name, ...)
}
\description{
#Quasi-Binomial Regression
#
#
}

