% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transformations.R
\name{ListToDataFrame}
\alias{ListToDataFrame}
\title{Converts a list of variable or data frames into a data.frame.}
\usage{
ListToDataFrame(list.of.variables, coerce.to.numeric = FALSE)
}
\arguments{
\item{list.of.variables}{A variable in a \code{DataSet} or \code{data.frame}.}

\item{coerce.to.numeric}{Makes factors and ordered factors numeric.}
}
\description{
Converts a list of variable or data frames into a data.frame.
}

