% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning.R
\name{MachineLearning}
\alias{MachineLearning}
\title{Train a machine learning model.}
\usage{
MachineLearning(algorithm, ..., warn.if.no.match = FALSE)
}
\arguments{
\item{algorithm}{The name of algorithm to perform the training. One of
\code{"CART"}, \code{"Regression"}, \code{"Random Forest"}, \code{"Deep Learning"},
\code{"Support Vector Machine"}, \code{"Linear Discriminant Analysis"}, or
\code{"Gradient Boosting"}.}

\item{...}{Arguments to the function \code{algorithm}; see See Also}

\item{warn.if.no.match}{Logical; If TRUE, a warning is shown if any arguments are not matched.}
}
\value{
A machine learning object.
}
\description{
Train a machine learning model.
}
\seealso{
\code{\link{SupportVectorMachine}}, \code{\link[flipTrees]{CART}},
\code{\link{DeepLearning}}, \code{\link{GradientBoosting}},
\code{\link{LinearDiscriminantAnalysis}}, \code{\link{RandomForest}},
\code{\link[flipRegression]{Regression}}
}
