% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{MachineLearningEnsemble}
\alias{MachineLearningEnsemble}
\title{Create an ensemble of MachineLearning and/or Regression models}
\usage{
MachineLearningEnsemble(models, compare.only = FALSE,
  evaluation.subset = NULL, evaluation.weights = NULL,
  output = "Comparison")
}
\arguments{
\item{models}{A \code{list} of models, all of which are of class
\code{MachineLearning} or \code{Regression}.}

\item{compare.only}{Logical; whether to just produce a table comparing the models or
additionally combine them to make a new ensemble model.}

\item{evaluation.subset}{An optional vector specifying a subset of observations to be
used for evaluating the models. If not specified, models will only be compared on the
training data. To evaluate on the whole sample, a subset must still be specified.}

\item{evaluation.weights}{An optional vector of weights to be used for evaluating the models.
Ignored if no evaluation.subset is supplied. A warning is given if these differ from the
training weights.}

\item{output}{If \code{compare.only} is \code{FALSE}, one of \code{"Comparison"} which
produces a table comparing the models, or \code{"Ensemble"} which produces a
\code{\link{ConfusionMatrix}}.}
}
\description{
Create an ensemble of MachineLearning and/or Regression models
}
