% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{MachineLearningMulti}
\alias{MachineLearningMulti}
\title{Create an ensemble or comparison table of new MachineLearning and/or Regression models}
\usage{
MachineLearningMulti(formula, data = NULL, subset = NULL,
  weights = NULL, evaluation.subset = NULL,
  missing = "Exclude cases with missing data", show.labels = FALSE,
  seed = 12321, models.args = NULL, compare.only = FALSE,
  optimal.ensemble = FALSE, output = "Comparison")
}
\arguments{
\item{formula}{A formula of the form \code{groups ~ x1 + x2 + ...}
That is, the response is the grouping factor and the right hand side
specifies the (non-factor) discriminators, and any transformations, interactions,
or other non-additive operators apart from \code{.} will be ignored.}

\item{data}{A \code{\link{data.frame}} from which variables specified
in formula are preferentially to be taken.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process, or, the name of a variable in \code{data}. It
may not be an expression.}

\item{weights}{An optional vector of sampling weights, or the
name of a variable in \code{data}. It may not be an expression.}

\item{evaluation.subset}{An optional vector specifying a subset of observations to be
used for evaluating the models. If not specified, models will only be compared on the
training data. If models are not trained on the whole sample To evaluate on the whole sample,
a subset must still be specified.}

\item{missing}{How missing data is to be treated. Options:
\code{"Error if missing data"},
\code{"Exclude cases with missing data"}, or
\code{"Imputation (replace missing values with estimates)"}.}

\item{show.labels}{Shows the variable labels, as opposed to the labels, in the outputs, where a
variables label is an attribute (e.g., attr(foo, "label")).}

\item{seed}{The random number seed.}

\item{models.args}{A \code{\link{list}} of lists of arguments to be passed to \code{\link{MachineLearning}}
to create each model.}

\item{compare.only}{Logical; whether to just produce a table comparing the models or
additionally combine them to make a new ensemble model.}

\item{optimal.ensemble}{Logical; whether to find the ensemble with the best accuracy or
r-squared, calculated on the \code{evaluation.subset} if given, else on the training
data. Ignored if \code{compare.only} is TRUE.}

\item{output}{If \code{compare.only} is \code{FALSE}, one of \code{"Comparison"} which
produces a table comparing the models, or \code{"Ensemble"} which produces a
\code{\link{ConfusionMatrix}}.}
}
\description{
Create an ensemble or comparison table of new MachineLearning and/or Regression models
}
