% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodnessoffit.R
\name{McFaddensRhoSquared}
\alias{McFaddensRhoSquared}
\title{\code{McFaddensRhoSquared}}
\usage{
McFaddensRhoSquared(x)
}
\arguments{
\item{model}{A 'Regression' model.}

\item{n.permutations}{Number of permutations used in computing the p-value.}
}
\description{
\code{McFaddensRhoSquared}
}
\details{
1 - the deviance divided by the null deviance.
 McFadden, D. (1974) “Conditional logit analysis of qualitative choice behavior.” Pp. 105-142 in P. Zarembka (ed.), Frontiers in Econometrics
}

