% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingdata.R
\name{MissingValuesByVariable}
\alias{MissingValuesByVariable}
\title{\code{MissingValuesByVariable}}
\usage{
MissingValuesByVariable(data)
}
\arguments{
\item{data}{A \code{\link{data.frame}}.}
}
\description{
Computes the number of missing values by variable.
}

