% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{Observed.Regression}
\alias{Observed.Regression}
\title{#' @export
Observed.default <- function(x)
{
    fitted(x)
}
#@method observed Regression}
\usage{
\method{Observed}{Regression}(x)
}
\description{
#' @export
Observed.default <- function(x)
{
    fitted(x)
}
#@method observed Regression
}

