% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regression.R
\name{OrderedLogit}
\alias{OrderedLogit}
\title{\code{OrderedLogit}
Ordered Logit Regression}
\usage{
OrderedLogit(formula, data, subset = NULL, weights = NULL, ...)
}
\description{
\code{OrderedLogit}
Ordered Logit Regression
}

