% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{OrderedToNumeric}
\alias{OrderedToNumeric}
\title{Converts an ordered factor into a numeric variable}
\usage{
OrderedToNumeric(ordered.factor)
}
\arguments{
\item{ordered.factor}{A variable in a \code{DataSet} or \code{data.frame}.}
}
\description{
Converts an ordered factor into a numeric variable
}

