% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning.R
\name{OrganizeCategoricalPredictors}
\alias{OrganizeCategoricalPredictors}
\title{OrganizeCategoricalPredictors}
\usage{
OrganizeCategoricalPredictors(input.model, all.combo.boxes)
}
\arguments{
\item{input.model}{The machine learning model.}

\item{all.combo.boxes}{The selection of predictor values.}
}
\description{
Organize information about categorical predictors in the model.
}
