% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressiondiagnostics.R
\name{OutlierTest}
\alias{OutlierTest}
\title{\code{OutlierTest}}
\usage{
OutlierTest(model)
}
\arguments{
\item{model}{A 'Regression' model.}
}
\description{
\code{OutlierTest}
}
\details{
Computes studentized residuals.
}

