% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{PoissonRegression}
\alias{PoissonRegression}
\title{Poisson Regression}
\usage{
PoissonRegression(formula, data, weights = NULL, subset = NULL, ...)
}
\arguments{
\item{...}{Additional arguments affecting the goodness-of-fit displayed.}
}
\description{
Poisson Regression
}

