% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{predict.DeepLearning}
\alias{predict.DeepLearning}
\title{\code{predict.DeepLearning}}
\usage{
\method{predict}{DeepLearning}(object, newdata = object$model,
  na.action = na.pass, ...)
}
\arguments{
\item{object}{A \code{DeepLearning} object.}

\item{newdata}{Optionally, a data frame in which to look for variables with which to predict.
If omitted, the data used to fit the model is used.}

\item{na.action}{Function determining what should be done with missing values in \code{newdata}.
The default is to predict \code{NA}.}

\item{...}{Additional arguments to pass to predict.DArch}
}
\description{
Predicts values for numeric outcomes and group membership for categories, for the entire sample passed into the original analysis (including missing and filtered values).
}

