% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{predict.GradientBoost}
\alias{predict.GradientBoost}
\title{\code{predict.GradientBoost}}
\usage{
\method{predict}{GradientBoost}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A \code{GradientBoost} object.}

\item{newdata}{Optionally, a data frame including the variables used to fit the model.
If omitted, the \code{data} supplied to \code{GradientBoost()} is used before any filtering.}

\item{...}{Additional arguments to pass to predict.xgb.Booster.}
}
\description{
Predicts values for numeric outcomes and group membership for categories based on \code{newdata}
and a fitted \code{GradientBoost} object.  A value (which may be NA) is returned for every instance
including those with missing data. If \code{newdata} is not specified the fitted \code{data} before filtering
are used.  NA is returned for cases with unfitted factor levels or any missing data.
}
