% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{predict.LDA}
\alias{predict.LDA}
\title{\code{predict.LDA}}
\usage{
\method{predict}{LDA}(object, newdata = object$model, na.action = na.pass,
  ...)
}
\arguments{
\item{object}{A \code{LDA} object.}
}
\description{
Predicts group membership  for the entire sample passed into the original analysis (including missing and filtered values).
}

