% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{predict.RandomForest}
\alias{predict.RandomForest}
\title{\code{predict.RandomForest}}
\usage{
\method{predict}{RandomForest}(object, newdata = object$model,
  na.action = na.pass, ...)
}
\arguments{
\item{object}{A \code{RandomForest} object.}
}
\description{
Predicts values for numeric outcomes and group membership for categories, for the entire sample passed into the original analysis (including missing and filtered values).
}

