% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{predict.SupportVectorMachine}
\alias{predict.SupportVectorMachine}
\title{\code{predict.SupportVectorMachine}}
\usage{
\method{predict}{SupportVectorMachine}(object, newdata = NULL,
  na.action = na.omit, ...)
}
\arguments{
\item{object}{A \code{SupportVectorMachine} object.}

\item{newdata}{Optionally, a data frame including the variables used to fit the model.
If omitted, the actual data used to fit the model is used (before any filtering).}

\item{na.action}{Function determining what should be done with missing values in \code{newdata}.
The default is to predict \code{NA}.}

\item{...}{Additional arguments to pass to predict.SupportVectorMachine.}
}
\description{
Predicts values for numeric outcomes and group membership for categories based on new data or the data
used to fit the model.  A value is returned for every instance including missing and filtered cases,
producing NA where no prediction is possible (eg missing data or unfitted factor levels).
}

