% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regressionplots.R
\name{PredictionPlot}
\alias{PredictionPlot}
\title{\code{Equation} Formats a model as an equation (string), for
description purposes.}
\usage{
PredictionPlot(Regression.object)
}
\arguments{
\item{Regression.object}{The model.}
}
\description{
\code{Equation} Formats a model as an equation (string), for
description purposes.
}

