% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning.R
\name{PredictOutcome}
\alias{PredictOutcome}
\title{PredictOutcome}
\usage{
PredictOutcome(input.model, DF, is.numeric)
}
\arguments{
\item{input.model}{The machine learning model.}

\item{DF}{The data frame containing the predictor values.}

\item{is.numeric}{A logical value indicating if the outcome is numeric.}
}
\description{
Predict the outcome for a simulator for a machine learning model.
}
