% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning.R
\name{PredictProbabilities}
\alias{PredictProbabilities}
\title{PredictProbabilities}
\usage{
PredictProbabilities(input.model, DF)
}
\arguments{
\item{input.model}{The machine learning model.}

\item{DF}{The data frame containing the predictor values.}
}
\description{
Predict the probabilities for a simulator for a machine learning model.
}
