% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/correspondenceanalysis.R
\name{print.CorrespondenceAnalysis}
\alias{print.CorrespondenceAnalysis}
\title{\code{print.CorrespondenceAnalysis}}
\usage{
\method{print}{CorrespondenceAnalysis}(CorrespondenceAnalysis.object)
}
\arguments{
\item{CorrespondenceAnalysis.object}{CorrespondenceAnalysis object.}
}
\description{
Creates a plot displaying the correspondence analysis results.
}

