% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeplearning.R
\name{print.DeepLearning}
\alias{print.DeepLearning}
\title{\code{print.DeepLearning}}
\usage{
\method{print}{DeepLearning}(x, ...)
}
\arguments{
\item{x}{DeepLearning object to print.}

\item{...}{Not used.}
}
\description{
\code{print.DeepLearning}
}

