% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineardiscriminantanalysis.R
\name{print.LDA}
\alias{print.LDA}
\title{print.LDA}
\usage{
\method{print}{LDA}(x, p.cutoff = 0.05, digits = max(3L,
  getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{The \link{LDA} object.}

\item{p.cutoff}{The alpha value to use when highlighting results.}

\item{digits}{The number of digits when printing the \code{"detail"} output.}

\item{...}{Generic print arguments.}
}
\description{
print.LDA
}
