% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{Probabilities.DeepLearning}
\alias{Probabilities.DeepLearning}
\title{\code{Probabilities.DeepLearning}}
\usage{
Probabilities.DeepLearning(x)
}
\arguments{
\item{x}{A \code{DeepLearning} object.}
}
\description{
Estimates probabilities of group membership for the entire sample passed into the original analysis (including missing and filtered values).
}

