% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{Probabilities.MachineLearningEnsemble}
\alias{Probabilities.MachineLearningEnsemble}
\title{\code{Probabilities.MachineLearningEnsemble}}
\usage{
\method{Probabilities}{MachineLearningEnsemble}(object)
}
\arguments{
\item{object}{A \code{MachineLearningEnsemble} object.}
}
\description{
Estimates probabilities of group membership for the entire sample passed into the original analysis (including
missing and filtered values).
}
