% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{Probabilities.RandomForest}
\alias{Probabilities.RandomForest}
\title{\code{Probabilities.RandomForest}}
\usage{
\method{Probabilities}{RandomForest}(x)
}
\arguments{
\item{x}{A \code{RandomForest} object.}
}
\description{
Estimates probabilities of group membership for the entire sample passed into the original analysis (including missing and filtered values).
}
