% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{Probabilities.SupportVectorMachine}
\alias{Probabilities.SupportVectorMachine}
\title{\code{Probabilities.SupportVectorMachine}}
\usage{
\method{Probabilities}{SupportVectorMachine}(x)
}
\arguments{
\item{x}{A \code{SupportVectorMachine} object.}
}
\description{
Estimates probabilities of group membership for the entire sample passed into the original analysis (including missing and filtered values).
}
