% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{PropensityWeights}
\alias{PropensityWeights}
\title{Compute the Propensity Weight scores for a binary classification model}
\usage{
PropensityWeights(object)
}
\arguments{
\item{object}{A \code{Regression} or \code{MachineLearning} output.}
}
\value{
A vector of numeric propensity weights for each case in the model.
}
\description{
Computes the propensity weights from a binary classification model
  using the inverse predicted probabilities of each observation being in the
  positive class.
}
