% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regression.R
\name{QuasiPoissonRegression}
\alias{QuasiPoissonRegression}
\title{\code{QuasiPoissonRegression}}
\usage{
QuasiPoissonRegression(formula, data, subset = NULL, weights = NULL, ...)
}
\description{
Quasi-Poisson Regression
}

