% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomforest.R
\name{RandomForest}
\alias{RandomForest}
\title{Fit a random forest model}
\usage{
RandomForest(formula, data = NULL, subset = NULL, weights = NULL,
  output = "Importance", missing = "Exclude cases with missing data",
  seed = 12321, show.labels = FALSE, sort.by.importance = TRUE, ...)
}
\arguments{
\item{formula}{A formula of the form \code{groups ~ x1 + x2 + ...}
That is, the response is the grouping factor and the right hand side
specifies the (non-factor) discriminators, and any transformations, interactions,
or other non-additive operators apart from \code{.} will be ignored.}

\item{data}{A \code{\link{data.frame}} from which variables specified
in formula are preferentially to be taken.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process, or the name of a variable in \code{data}. It
may not be an expression.}

\item{weights}{An optional vector of sampling weights, or the
name of a variable in \code{data}. It may not be an expression.}

\item{output}{One of \code{"Importance"}, \code{"Prediction-Accuracy Table"} or \code{"Detail"}.}

\item{missing}{How missing data is to be treated. Options:
\code{"Error if missing data"},
\code{"Exclude cases with missing data"}, or
\code{"Imputation (replace missing values with estimates)"}.}

\item{seed}{The random number seed.}

\item{show.labels}{Shows the variable labels, as opposed to the labels, in the outputs, where a
variables label is an attribute (e.g., attr(foo, "label")).}

\item{sort.by.importance}{Sort the last column of the importance table
in descending order.}

\item{...}{Other arguments to be supplied to \code{\link{randomForest}}.}
}
\description{
Fit a random forest model
}
