% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractchartdata.R, R/variables.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ExtractChartData}
\alias{Probabilities}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{flipData}{\code{\link[flipData]{Probabilities}}}

  \item{flipFormat}{\code{\link[flipFormat]{ExtractChartData}}}
}}

