% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/missingdata.R
\name{SampleDescription}
\alias{SampleDescription}
\title{\code{SampleDescription}}
\usage{
SampleDescription(n.total, n.subset, n.estimation, subset.label,
  weighted = TRUE, weight.label = "", missing)
}
\arguments{
\item{n.total}{Total number of observations in the database.}

\item{n.subset}{Total number of observations in the subset (less than or equal to \code{n.total}).}

\item{n.estimation}{The total number of observations used in estimation (less than or equal to \code{subset}).}

\item{subset.label}{Total number of observations in the database.}

\item{weighted}{Total number of observations in the database.}

\item{weight.label}{Total number of observations in the database.}

\item{missing}{How missing data is to be treated in the analysis. Options are:
\code{"Error if missing data"}, \code{"Exclude cases with missing data"},
,and \code{"Imputation (replace missing values with estimates)"}.}
}
\description{
Describes the sample, for use as footers in multivariate analyses.
}

