% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/missingdata.R
\name{SingleImputation}
\alias{SingleImputation}
\title{\code{SingleImputation}}
\usage{
SingleImputation(formula, data, ...)
}
\arguments{
\item{formula}{A \code{\link{formula}}.}

\item{data}{A \code{\link{data.frame}}.}

\item{...}{Optional arguments for \code{\link{mice}}.
#}
}
\description{
This is a wrapper for \code{\link{mice}}.
}
\examples{
# df <- data.frame(x = c(NA, 1), y = 1:2)
# ExcludeCasesWithAnyMissingData(df)
}

