% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/missingdata.R
\name{SingleImputaton}
\alias{SingleImputaton}
\title{\code{SingleImputation}}
\usage{
SingleImputaton(data, outcome.name = NULL, ...)
}
\arguments{
\item{data}{A \code{\link{data.frame}}.}

\item{outcome.name}{A If the name of the outcome (outcome)
variable is provided, any cases with missing values on this variable
are excluded from the final data file.}

\item{...}{Optional arguments for \code{\link{mice}}.
#}
}
\description{
This is a wrapper for \code{\link{mice}}.
}
\examples{
# df <- data.frame(x = c(NA, 1), y = 1:2)
# ExcludeCasesWithAnyMissingData(df)
}

