% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cart.R
\name{treeFrameToList}
\alias{treeFrameToList}
\title{\code{treeFrameToList} Converts a \code{\link{tree}} into a format usable in a sankeytree.}
\usage{
treeFrameToList(tree, max.tooltip.length = 150, show.whole.factor = FALSE,
  numeric.distribution = FALSE, num.color.div = 5, const.bin.size = TRUE)
}
\arguments{
\item{max.tooltip.length}{The maximum length of the tooltip (determines the scale of the tree).}

\item{show.whole.factor}{Controls whether or not all the factor levels are displayed in the tooltip.}

\item{numeric.distribution}{Outputs additional diagnostics in the tooltip.}

\item{num.color.div}{??.}

\item{const.bin.size}{??.}
}
\description{
\code{treeFrameToList} Converts a \code{\link{tree}} into a format usable in a sankeytree.
}

