% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressiondiagnostics.R
\name{UnusualObservations}
\alias{UnusualObservations}
\title{\code{UnusualObservations}}
\usage{
UnusualObservations(model)
}
\arguments{
\item{model}{A 'Regression'  model.}
}
\description{
\code{UnusualObservations}
}
\details{
Computes studentized residuals, Cook's distance, and hat values, and reports if any are above the threshhold values.
}

