% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeplearning.R
\name{VariableImportance}
\alias{VariableImportance}
\title{\code{VariableImportance}}
\usage{
VariableImportance(object)
}
\arguments{
\item{object}{Object of class \code{"DeepLearning"}}
}
\description{
Computes relative importance of input variable to the neural network
as the sum of the product of raw input-hidden, hidden-output connection weights
as proposed by Olden et al. 2004.
}

