% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/factoranalysis.R
\name{weightedPartialCovarianceMatrix}
\alias{weightedPartialCovarianceMatrix}
\title{#Q23 and Weights from Phone.sav
#
#25 variables from a 5-point scale. Extra missing data has been added at random.
#This makes up about 20% of the values. This is to test PCA and Factor analysis.
#
#}
\format{A list containing:
#\describe{
#  \item{data.set}{25 variables from q23}
#  \item{weight}{A vector of weights}
#}}
\usage{
weightedPartialCovarianceMatrix(data, weight, correlation = FALSE)
}
\description{
#Q23 and Weights from Phone.sav
#
#25 variables from a 5-point scale. Extra missing data has been added at random.
#This makes up about 20% of the values. This is to test PCA and Factor analysis.
#
#
}

