% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/weightedcomputations.R
\name{WeightedSVD}
\alias{WeightedSVD}
\title{Weighted Singular Value Decomposition
\code{WeightedSVD} Computes a SVD with frequency weights.}
\usage{
WeightedSVD(x, weights = rep(1, n), nu = min(n, p))
}
\arguments{
\item{x}{A numeric or complex matrix whose SVD decomposition is to be computed. Logical matrices are coerced to numeric.}

\item{weights}{Frequency Weights.}

\item{nu}{the number of left singular vectors to be computed. This must between 0 and n = nrow(x).}
}
\description{
Weighted Singular Value Decomposition
\code{WeightedSVD} Computes a SVD with frequency weights.
}

