% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regression.R
\name{BinaryLogit}
\alias{BinaryLogit}
\title{\code{BinaryLogit} Binary Logit Regression.}
\usage{
BinaryLogit(formula, data, subset = NULL, weights = NULL, ...)
}
\description{
\code{BinaryLogit} Binary Logit Regression.
}

