% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regressiondiagnostics.R
\name{BreuschPagan}
\alias{BreuschPagan}
\title{\code{BreuschPagan}Breusch-Pagan test for non-constant variance.}
\usage{
BreuschPagan(Regression.object)
}
\arguments{
\item{Regression.object}{An object of class \code{\link{Regression}}.}
}
\description{
\code{BreuschPagan}Breusch-Pagan test for non-constant variance.
}
\details{
Weights are ignored when conducting the test.
}

