% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cart.R
\name{CART}
\alias{CART}
\title{\code{CART} Creats a classification or regression tree.}
\usage{
CART(formula, data, weights = NULL, subset = NULL, output = "Sankey Plot",
  ...)
}
\arguments{
\item{formula}{A formula expression. The left-hand-side (response) should
be either a numerical vector when a regression tree will be fitted or
a factor, when a classification tree is produced. The right-hand-side should
be a series of numeric or factor variables separated by \code{+}; there should be
no interaction terms. Both \code{.} and \code{-} are allowed: regression trees can have
offset terms}

\item{data}{A data frame in which to preferentially interpret formula, weights and subset}

\item{weights}{An optional vector of sampling weights, or, the name or,
the name of a variable in \code{data}. It may not be an expression.}

\item{subset}{An optional vector specifying a subset of observations to be used in the fitting process, or,
the name of a variable in \code{data}. It may not be an expression.
\code{subset} may not}

\item{output}{How the tree is represented: \code{"Sankey Plot"}, \code{"Tree"}, or \code{"Text"}.}

\item{...}{Additional arguments that are passed to  \code{\link{tree}}
and \code{\link{tree.control}}. Normally used for mincut, minsize or mindev}
}
\description{
\code{CART} Creats a classification or regression tree.
}
\details{
Creates a \code{\link{tree}} and plots it as a \code{\link{sankeytree}}
}

