% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeplearning.R
\name{DeepLearning}
\alias{DeepLearning}
\title{Fit a neural network model}
\usage{
DeepLearning(formula, data = NULL, subset = NULL, weights = NULL,
  output = "Accuracy", missing = "Exclude cases with missing data",
  seed = 12321, show.labels = FALSE, hidden.nodes, iterations = 100)
}
\arguments{
\item{formula}{A formula of the form \code{groups ~ x1 + x2 + ...}
That is, the response is the grouping factor and the right hand side
specifies the (non-factor) discriminators, and any transformations, interactions,
or other non-additive operators apart from \code{.} will be ignored.}

\item{data}{A \code{\link{data.frame}} from which variables specified
in formula are preferentially to be taken.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process, or, the name of a variable in \code{data}. It
may not be an expression.}

\item{weights}{An optional vector of sampling weights, or the
name of a variable in \code{data}. It may not be an expression.}

\item{output}{One of \code{"Accuracy"}, \code{"Prediction-Accuracy Table"}, \code{"Cross Validation"}
or \code{"Network Layers"}.}

\item{missing}{How missing data is to be treated. Options:
\code{"Error if missing data"},
\code{"Exclude cases with missing data"},}

\item{seed}{The random number seed used in imputation.}

\item{show.labels}{Shows the variable labels, as opposed to the labels, in the outputs, where a
variables label is an attribute (e.g., attr(foo, "label")).}

\item{hidden.nodes}{Numeric; a \code{\link{vector}} specfy the number of hidden nodes in each
hideen layer.}

\item{iterations}{Integer; the number of iterations to train the network.}
}
\description{
Fit a neural network model
}
