% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeplearning.R
\name{DeepLearning}
\alias{DeepLearning}
\title{\code{DeepLearning}}
\usage{
DeepLearning(formula, hidden = c(50, 50),
  unit.function = "rectifiedLinearUnit", learning.rate = 0.01,
  learning.decay = 0, dropout.input = 0, dropout.hidden = 0.5,
  dither = F, epochs = 100, batch.size = 1, bootstrap = T,
  pretrain.epoch = 0, data = NULL, subset = NULL, weights = NULL,
  output = "Summary", missing = "Exclude cases with missing data",
  seed = 12321, show.labels = FALSE, ...)
}
\arguments{
\item{formula}{A formula of the form \code{groups ~ x1 + x2 + ...}
That is, the response is the grouping factor and the right hand side
specifies the (non-factor) discriminators, and any transformations, interactions,
or other non-additive operators will be ignored.
transformations nor}

\item{hidden}{A vector of integers giving the number of units in each layer.}

\item{unit.function}{Type of function to use. By default it is \code{"rectifiedLinearUnit"}.}

\item{learning.rate, learning.decay, dropout.input, dropout.hidden, dither, bootstrap, pretrain.epoch}{Parameter passed to DArch.}

\item{epochs}{Duration of training period.}

\item{batch.size}{Size of training set}

\item{data}{A \code{\link{data.frame}} from which variables specified
in formula are preferentially to be taken.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process, or, the name of a variable in \code{data}. It
may not be an expression. \code{subset} may not}

\item{weights}{An optional vector of sampling weights, or, the name or, the
name of a variable in \code{data}. It may not be an expression.}

\item{output}{One of \code{"Importance"}, or \code{"Detail"}.}

\item{missing}{How missing data is to be treated in the regression. Options:
\code{"Error if missing data"},
\code{"Exclude cases with missing data"},}

\item{seed}{The random number seed used in imputation.}

\item{show.labels}{Shows the variable labels, as opposed to the labels, in the outputs, where a
variables label is an attribute (e.g., attr(foo, "label")).}

\item{...}{Other arguments to be supplied to \code{\link{darch}}.}
}
\description{
\code{DeepLearning}
}

