% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transformations.R
\name{DichotomizeFactor}
\alias{DichotomizeFactor}
\title{\code{DichotomizeFactor} Converts a list of variable or data frames into a data.frame.}
\usage{
DichotomizeFactor(variable, cutoff = 0.5, warning = FALSE,
  variable.name = deparse(substitute(variable)))
}
\arguments{
\item{list.of.variables}{A variable in a DataSet or data.frame.}

\item{coerce.to.numeric}{Makes factors and ordered factors numeric.}
}
\description{
\code{DichotomizeFactor} Converts a list of variable or data frames into a data.frame.
}

