% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{Equation}
\alias{Equation}
\title{Formats a model as an equation (string), for
description purposes.}
\usage{
Equation(object)
}
\arguments{
\item{object}{The model.}
}
\description{
Formats a model as an equation (string), for
description purposes.
}

