% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatting.R
\name{Equation}
\alias{Equation}
\title{Formats a model as an equation (string), for
description purposes. It shows the linear expectation of the mean, and assumes you have not manipulated
default assumptions (e.g., via offsets and link functions).}
\usage{
Equation(object)
}
\arguments{
\item{object}{The model.}
}
\description{
Formats a model as an equation (string), for
description purposes. It shows the linear expectation of the mean, and assumes you have not manipulated
default assumptions (e.g., via offsets and link functions).
}

