% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{FormatAsPercent}
\alias{FormatAsPercent}
\title{Converts a proportion to a percent and
formats it nicely and with a percentage sign.}
\usage{
FormatAsPercent(x, digits = 2, format = "fg", suffix = "\%", ...)
}
\arguments{
\item{x}{The number(s)}

\item{digits}{Number of significant digits}

\item{format}{See \code{\link{formatC}}.}

\item{suffix}{The text to appear after a percent. Defaults to \%.}

\item{...}{Additional arguments that are passed to \code{\link{formatC}}.}
}
\description{
Converts a proportion to a percent and
formats it nicely and with a percentage sign.
}

