% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatting.R
\name{FormatAsPValue}
\alias{FormatAsPValue}
\title{\code{FormatAsPValue}
Formats p relatively nicely, ensuring that values greater than 0.05
are not rounded to 0.05.}
\usage{
FormatAsPValue(p, p.cutoff = 0.05, ...)
}
\arguments{
\item{p}{The number(s)}

\item{...}{Additional arguments that are passed to \code{\link{formatC}}.}

\item{digits}{Number of significant digits}

\item{format}{See \code{\link{formatC}}.}
}
\description{
\code{FormatAsPValue}
Formats p relatively nicely, ensuring that values greater than 0.05
are not rounded to 0.05.
}

