% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{FormatAsReal}
\alias{FormatAsReal}
\title{Formats real numbers nicely, as if they are to be presented as an average.}
\usage{
FormatAsReal(x, digits = 2, format = "fg", ...)
}
\arguments{
\item{x}{The number(s)}

\item{digits}{Number of significant digits}

\item{format}{See \code{\link{formatC}}.}

\item{...}{Additional arguments that are passed to \code{\link{formatC}}.}
}
\description{
Formats real numbers nicely, as if they are to be presented as an average.
}

