% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regression.R
\name{LinearRegression}
\alias{LinearRegression}
\title{\code{LinearRegression}Linear Regression.}
\usage{
LinearRegression(formula, data, subset = NULL, weights = NULL,
  robust.se = FALSE, ...)
}
\arguments{
\item{formula}{An object of class \code{\link{formula}} (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under ‘Details’.}

\item{data}{A \code{\link{data.frame}}.}

\item{subset}{An optional vector specifying a subset of observations to be used in the fitting process.}

\item{weights}{An optional vector of sampling weights.}

\item{robust.se}{Computes standard errors that are robust to violations of the assumption of constant variance, using the HC3 modification of White's estimator.}

\item{...}{Additional argments to be past to  \code{\link{lm}} or, if the data
is weighted,  \code{\link{survey::svyglm}}.}
}
\description{
Linear regression: ordinary least squares or least squares with survey weights.
}

