% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regression.R
\name{LinearRegression}
\alias{LinearRegression}
\title{\code{LinearRegression}Linear Regression.}
\usage{
LinearRegression(formula, data, subset = NULL, weights = NULL,
  missing = "Exclude cases with missing data", robust.se = FALSE, ...)
}
\arguments{
\item{formula}{An object of class \code{\link{formula}} (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under ‘Details’.}

\item{data}{A \code{\link{data.frame}}.}

\item{subset}{An optional vector specifying a subset of observations to be used in the fitting process.}

\item{weights}{An optional vector of sampling weights.}

\item{missing}{How missing data is to be treated in the regression. Options are:
\code{"Error if missing data"}, \code{"Exclude cases with missing data"},
\code{"Use partial data (pairwise)"},and \code{"Imputation"}.}

\item{robust.se}{Computes standard errors that are robust to violations of the assumption of constant variance, using the HC3 modification of White's estimator.}

\item{...}{Additional argments to be past to  \code{\link{lm}} or, if the data
is weighted,  \code{\link{survey::svyglm}}.}
}
\description{
Linear regression: ordinary least squares or least squares with survey weights.
}
\details{
"Imputation" is performed using \code{\link{mice}}. All selected
outcome and predictor variables are included in the imputation. Then,
cases with missing values in the outcome variable are excluded from the
analysis (von Hippel 2007). Where "Use partial data (pairwise)" is used, if the data is weighted, a
synthetic data file is created by sampling with replacement in proportion to the weights,where the
sample size is the sum of the weights.
}
\references{
von Hippel, Paul T. 2007. "Regression With Missing Y's: An
Improved Strategy for Analyzing Multiply Imputed Data." Sociological Methodology 37:83-117.
}

