% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regression.R
\name{LinearRegression}
\alias{LinearRegression}
\title{\code{Linear Regression} Linear Regression.}
\usage{
LinearRegression(formula, data, weights = NULL, subset = NULL, ...)
}
\arguments{
\item{...}{Additional arguments affecting the goodness-of-fit displayed.}

\item{object}{An object for which a summary is desired..}
}
\description{
reports the goodness-of-fit of an object.
}
\details{
Uses \code{\link{Zelig:zelimethods}}
which depend on the \code{\link{class}} of the first argument.
}

