% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning.R
\name{MachineLearning}
\alias{MachineLearning}
\title{Train a machine learning model.}
\usage{
MachineLearning(algorithm, ..., warn.if.no.match = FALSE)
}
\arguments{
\item{algorithm}{The name of algorithm to perform the training.}

\item{...}{Arguments to the function \code{algorithm}}

\item{warn.if.no.match}{Logical; If TRUE, a warning is shown if any arguments are not matched.}
}
\value{
A machine learning object.
}
\description{
Train a machine learning model.
}
