% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{OrderedLogit}
\alias{OrderedLogit}
\title{Ordered Logistic Regression}
\usage{
OrderedLogit(formula, data, subset = NULL, weights = NULL, ...)
}
\arguments{
\item{...}{Additional arguments affecting the goodness-of-fit displayed.}
}
\description{
Ordered Logistic Regression
}

