% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regression.R
\name{OrderedLogit}
\alias{OrderedLogit}
\title{\code{OrderedLogit}
Ordered Logit Regression}
\usage{
OrderedLogit(formula, data, subset = NULL, weights = NULL, ...)
}
\arguments{
\item{formula}{An object of class \code{\link{formula}} (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under ‘Details’.}

\item{data}{A \code{\link{data.frame}}.}

\item{subset}{An optional vector specifying a subset of observations to be used in the fitting process.}

\item{weights}{An optional vector of sampling weights.}

\item{...}{Additional argments to be past to  \code{\link{lm}} or, if the data
is weighted,  \code{\link{survey::svyglm}}.}
}
\description{
\code{OrderedLogit}
Ordered Logit Regression
}

