% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transformations.R
\name{OrderedToNumeric}
\alias{OrderedToNumeric}
\title{\code{OrderedToNumeric} Converts an ordered factor into a numeric variable}
\usage{
OrderedToNumeric(ordered.factor)
}
\arguments{
\item{variable}{A variable in a DataSet or data.frame.}

\item{variable.name}{The name of the variable.}
}
\description{
\code{OrderedToNumeric} Converts an ordered factor into a numeric variable
}

