% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeplearning.R
\name{predict.DeepLearning}
\alias{predict.DeepLearning}
\title{\code{predict.DeepLearning}}
\usage{
\method{predict}{DeepLearning}(object, newdata = NULL)
}
\arguments{
\item{object}{Object of class \code{"DeepLearning"}.}

\item{newdata}{Data to which to apply the prediction}

\item{...}{Not used}
}
\description{
\code{predict.DeepLearning}
}

