% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeplearning.R
\name{predict.DeepLearning}
\alias{predict.DeepLearning}
\title{\code{predict.DeepLearning}}
\usage{
\method{predict}{DeepLearning}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A \code{DeepLearning} object.}

\item{newdata}{Optionally, a data frame including the variables used to fit the model.
If omitted, the \code{data} supplied to \code{DeepLearning()} is used before any filtering.}

\item{...}{Additional arguments to pass to predict.DeepLearning.}
}
\description{
Predicts values for numeric outcomes and group membership for categories based on \code{newdata}
and a fitted DeepLearning \code{object}.  A value (which may be NA) is returned for every instance
including those with missing data and for the fitted \code{data} before filtering in the case
that \code{newdata} is not specified.  NA is returned for cases with unfitted factor levels.
}
