% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{predict.LDA}
\alias{predict.LDA}
\title{\code{predict.LDA}}
\usage{
\method{predict}{LDA}(object, newdata = NULL, na.action = na.pass, ...)
}
\arguments{
\item{object}{A \code{LDA} object.}

\item{newdata}{Optionally, a data frame including the variables used to fit the model.
If omitted, the \code{data} supplied to \code{LDA()} is used before any filtering.}

\item{na.action}{Function determining what should be done with missing values in \code{newdata}.
The default is to predict \code{NA}.}

\item{...}{Additional arguments to pass to predict.lda.}
}
\description{
Predicts a model outcome based on \code{newdata} and a fitted LDA \code{object}.  A value (which
may be NA) is returned for every instance including those with missing data and for the
fitted \code{data} before filtering in the case that \code{newdata} is not specified.
NA is returned for cases with unfitted factor levels.
}
