% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regressionplots.R
\name{PredictionPlot}
\alias{PredictionPlot}
\title{Plots the observed, predict and an independent variable from a
regression model.}
\usage{
PredictionPlot(Regression.object, predictor.number = NULL)
}
\arguments{
\item{Regression.object}{The model.}

\item{predictor.number}{The index of the predictor to be used in the plot,
where 1 means the first predictor.}
}
\description{
Plots the observed, predict and an independent variable from a
regression model.
}
\details{
With numeric predicts, other variables are assumed to have an
average effect. With factors, the first level is from the intercept, and
the other levels are the intercept plus the parameter (thus, if you modify
the contrats, this plot will not display the predicted values correctly).
}

