% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeplearning.R
\name{Probabilities.DeepLearning}
\alias{Probabilities.DeepLearning}
\title{\code{Probabilities.DeepLearning}}
\usage{
\method{Probabilities}{DeepLearning}(object)
}
\arguments{
\item{object}{A \code{DeepLearning} object.}
}
\description{
Estimates probabilities of class membership for the entire sample passed into the original
analysis (including missing and filtered values).
}
