% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{Probabilities.GradientBoost}
\alias{Probabilities.GradientBoost}
\title{\code{Probabilities.GradientBoost}}
\usage{
\method{Probabilities}{GradientBoost}(object)
}
\arguments{
\item{object}{A \code{GradientBoost} object.}
}
\description{
Estimates probabilities of group membership for the entire sample passed into the original analysis (including
missing and filtered values).
}
