% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{QuasiPoissonRegression}
\alias{QuasiPoissonRegression}
\title{Quasi-Poisson Regression}
\usage{
QuasiPoissonRegression(formula, data, weights = NULL, subset = NULL, ...)
}
\arguments{
\item{...}{Additional arguments affecting the goodness-of-fit displayed.}
}
\description{
Quasi-Poisson Regression
}

