% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{Regression}
\alias{Regression}
\title{\code{Regression} Generalized Regression.}
\usage{
Regression(formula, data = NULL, subset = NULL, weights = NULL,
  missing = "Exclude cases with missing data", type = "Linear",
  robust.se = FALSE, method = "default", detail = TRUE, ...)
}
\arguments{
\item{formula}{An object of class \code{\link{formula}} (or one that can be
coerced to that class): a symbolic description of the model to be fitted.
The details of type specification are given under \sQuote{Details}.}

\item{data}{A \code{\link{data.frame}}.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process, or, the name of a variable in \code{data}. It
may not be an expression. \code{subset} may not}

\item{weights}{An optional vector of sampling weights, or, the name or, the
name of a variable in \code{data}. It may not be an expression.}

\item{missing}{How missing data is to be treated in the regression. Options
are: \code{"Error if missing data"}, \code{"Exclude cases with missing
data"}, \code{"Use partial data (pairwise correlations)"},and
\code{"Imputation (replace missing values with estimates)"}.}

\item{type}{Defaults to \code{"linear"}. Other types are: \code{"Poisson"},
\code{"Quasi-Poisson"}, \code{"Binary Logit"}, \code{"NBD"},
\code{"Ordered Logit"}, and \code{"Multinomial Logit"}}

\item{robust.se}{Computes standard errors that are robust to violations of
the assumption of constant variance, using the HC1 (degrees of freedom)
modification of White's (1980) estimator (Long and Ervin, 2000).}

\item{method}{The method to be used; for fitting. This will only do something if
method = "model.frame", which returns the model frame.}

\item{detail}{More detailed outputs.}

\item{...}{Additional argments to be past to  \code{\link{lm}} or, if the
data is weighted,  \code{\link[survey]{svyglm}}.}
}
\description{
\code{Regression} Generalized Regression.
}
\details{
"Imputation (replace missing values with estimates)". All selected
  outcome and predictor variables are included in the imputation, excluding
  cases that are excluded via subset or have invalid weights, but including
  cases with missing values of the outcome variable.
  Then, cases with missing values in the outcome variable are excluded from
  the analysis (von Hippel 2007). Where "Use partial data (pairwise
  correlations)" is used, if the data is weighted, a synthetic data file is
  created by sampling with replacement in proportion to the weights,where the
  sample size is the sum of the weights. See \code{\link{SingleImputation}}.
}
\references{
von Hippel, Paul T. 2007. "Regression With Missing Y's: An
  Improved Strategy for Analyzing Multiply Imputed Data." Sociological
  Methodology 37:83-117. White, H. (1980), A heteroskedastic-consistent
  covariance matrix estimator and a direct test of heteroskedasticity.
  Econometrica, 48, 817-838. Long, J. S. and Ervin, L. H. (2000). Using
  heteroscedasticity consistent standard errors in the linear regression
  model. The American Statistician, 54( 3): 217-224.
}

