% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportvectormachine.R
\name{SupportVectorMachine}
\alias{SupportVectorMachine}
\title{Fit a support vector machine model}
\usage{
SupportVectorMachine(formula, data = NULL, subset = NULL,
  weights = NULL, output = "Accuracy",
  missing = "Exclude cases with missing data", cost = 1,
  seed = 12321, statistical.assumptions, show.labels = FALSE, ...)
}
\arguments{
\item{formula}{A formula of the form \code{groups ~ x1 + x2 + ...}
That is, the response is the grouping factor and the right hand side
specifies the (non-factor) discriminators, and any transformations, interactions,
or other non-additive operators apart from \code{.} will be ignored.}

\item{data}{A \code{\link{data.frame}} from which variables specified
in formula are preferentially to be taken.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process, or, the name of a variable in \code{data}. It
may not be an expression.}

\item{weights}{An optional vector of sampling weights, or the
name of a variable in \code{data}. It may not be an expression.}

\item{output}{One of \code{"Accuracy"}, \code{"Prediction-Accuracy Table"} or \code{"Detail"}.}

\item{missing}{How missing data is to be treated. Options:
\code{"Error if missing data"},
\code{"Exclude cases with missing data"}, or
\code{"Imputation (replace missing values with estimates)"}.}

\item{cost}{A positive number controlling the compromoise between exactly fitting the training data
(higher cost) and the ability to generalise to unseen data (lower cost).}

\item{seed}{The random number seed.}

\item{statistical.assumptions}{A Statistical Assumptions object.}

\item{show.labels}{Shows the variable labels, as opposed to the labels, in the outputs, where a
variables label is an attribute (e.g., attr(foo, "label")).}

\item{...}{Other arguments to be supplied to \code{\link{svm}}.}
}
\description{
Fit a support vector machine model
}
