% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{treeFrameToList}
\alias{treeFrameToList}
\title{\code{treeFrameToList} Converts a \code{\link{tree}} into a format usable in
a sankeytree.}
\usage{
treeFrameToList(tree, max.tooltip.length = 150, show.whole.factor = FALSE,
  numeric.distribution = FALSE, custom.color = TRUE, num.color.div = 101,
  const.bin.size = TRUE, draw.legend = TRUE)
}
\arguments{
\item{tree}{The tree to convert.}

\item{max.tooltip.length}{The maximum length of the tooltip (determines the
scale of the tree).}

\item{show.whole.factor}{Controls whether or not all the factor levels are
displayed in the tooltip.}

\item{numeric.distribution}{Outputs additional diagnostics in the tooltip.}

\item{custom.color}{logical; if \code{true}, generates custom tree color,
else use colors provided by sankeyTree package.}

\item{num.color.div}{positive integer in the range [2,inf]. Controls the
color resolution of the tree. A higher value gives a smoother color
transition.}

\item{const.bin.size}{logical; if \code{true}, each color spans an equal step
of y-value or an equal number of points.}

\item{draw.legend}{logical; if \code{true}, output the colors as a sorted RBG
value list to draw legend}
}
\description{
\code{treeFrameToList} Converts a \code{\link{tree}} into a format usable in
a sankeytree.
}

